src/cuve-api/requirements.txt
fastapi==0.115.6
httpx==0.28.1
jinja2==3.1.4
pytest==8.3.4
python-dotenv==1.0.1
uvicorn[standard]==0.34.0
